add docs for guarded and guardedvec.
