+++ client/src/utils/ids.rs
@@ -218,38 +218,3 @@ impl Into<String> for VaultId {
         self.0.as_ref().base64()
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_client_id() {
-        let path = b"some_path";
-        let data = b"a bunch of random data";
-        let mut buf = [0; 64];
-
-        let id = ClientId::load_from_path(data, path).unwrap();
-
-        HMAC_SHA512(data, path, &mut buf);
-
-        let (test, _) = buf.split_at(24);
-
-        assert_eq!(ClientId::load(test).unwrap(), id);
-    }
-
-    #[test]
-    fn test_vault_id() {
-        let path = b"another_path_of_data";
-        let data = b"a long sentance for seeding the id with some data and bytes.  Testing to see how long this can be without breaking the hmac";
-        let mut buf = [0; 64];
-
-        let id = VaultId::load_from_path(data, path).unwrap();
-
-        HMAC_SHA512(data, path, &mut buf);
-
-        let (test, _) = buf.split_at(24);
-
-        assert_eq!(VaultId::load(test).unwrap(), id);
-    }
-}
