+++ engine/runtime/src/guarded.rs
@@ -8,6 +8,25 @@ use core::{
     ops::{Deref, DerefMut},
 };
 
+/// A guarded type for protecting fixed-length secrets allocated on the heap.
+///
+/// Provides the following features and guarantees:
+/// * Causes segfault upon access without using a borrow.
+/// * Protected using mprotect:
+///    * `Prot::NoAccess` - when the box has no current borrows.
+///    * `Prot::ReadOnly` - when the box has at least one current immutable borrow.
+///    * `Prot::ReadWrite` - when the box has a current mutable borrow (can only have one at a time).
+/// * The allocated memory uses guard pages both proceeding and following the memory. Overflows and large underflows
+///   cause immediate termination of the program.
+/// * A canary proceeds the memory location to detect smaller underflows.  The program will drop the underlying memory
+///   and terminate if detected.
+/// * The Memory is locked with `mlock`.
+/// * When the memory is freed, `munlock` is called.
+/// * The memory is zeroed when no longer in use via `sodium_free`.
+/// * `Guarded` types can be compared in constant time.
+/// * `Guarded` types can not be printed using `Debug`.
+/// * The interior data of a `Guarded` type may not be `Clone`.
+
 #[derive(Clone, Eq)]
 pub struct Guarded<T: Bytes> {
     boxed: Boxed<T>,
@@ -145,13 +164,13 @@ impl<T: Bytes + Randomized> Guarded<T> {
     }
 }
 
-impl<T: Bytes + ZeroOut> Guarded<T> {
+impl<T: Bytes + Zeroed> Guarded<T> {
     pub fn zero() -> Self {
         Self { boxed: Boxed::zero(1) }
     }
 }
 
-impl<T: Bytes + ZeroOut> From<&mut T> for Guarded<T> {
+impl<T: Bytes + Zeroed> From<&mut T> for Guarded<T> {
     fn from(data: &mut T) -> Self {
         Self { boxed: data.into() }
     }
